Refactor context_data
