+++ src/model.rs
@@ -1,10 +1,6 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::{collections::{BTreeMap}, fmt::{self, Display}};
-use k8s_openapi::ByteString;
-use crate::constants;
-
-
 
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema, Copy)]
